Move censor event functions to separate class
